+++ app/src/main/java/com/xiaolian/amigo/data/network/model/wallet/RechargeDenominations.java
@@ -9,8 +9,8 @@ import lombok.Data;
  */
 @Data
 public class RechargeDenominations {
-    private int activityType;
-    private int amount;
-    private int id;
-    private int value;
+    private Integer activityType;
+    private Long amount;
+    private Long id;
+    private Long value;
 }
